Added 2 cameras. 1 light to scene, fixed some bugs
Added two news cameras to our test scene, spf.xml. The first one is a perspective
camera from the UFO, ponting to the entrance of the power plant, and the other
is an orthogonal camera from the high x direction, giving a front view of the
plant.
The added light is a spotlight from near the origin, towards the plant.

Fixed two things. The first one was a bug where the degree angle parameter in
the scene files wouldn't be converted to radian, which skewed the perspective
angles very fast.
The second one was a quality of light improvement. The settings interface now
opens up automatically at startup.